# apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Mesh gateway
  version: 0.1.0
  summary: Trading fees become hourly AI inference credits, spent through an OpenAI-compatible gateway, served by Macs and frontier providers, and sold on a marketplace when unused.
  description: |
    Mesh is a token whose trading fees buy AI inference for the people holding it. A 1.5 % fee on every
    $MESH trade is swept once an hour; half goes to the network treasury and half is split, pro rata, across
    every wallet holding at least 1,000 tokens (time-weighted over the hour) as credits denominated in US
    dollars. A second engine, a share of the margin on paid requests and marketplace fees, pays into the same
    hourly pool (`GET /stats → usageShareEnabled`).
    Holders spend credits through this gateway with their own API keys: open models on Macs in the network at
    a flat price per million tokens, frontier models through zero-data-retention providers at list plus a 6 %
    markup that covers what the upstream charges Mesh (`GET /v1/models` carries both prices). Credits nobody
    will use are sold on the credit marketplace (`/market/*`, docs/MARKETPLACE.md), and anyone can buy credits
    from Mesh at face value with a prepaid balance (`/me/credits/buy`). A holder's first sign-in receives
    starter credits, which can be spent but not sold.

    What stands behind a credit, and what limits it:

    * **Reserve.** The holder share of every sweep is swapped to a stablecoin on chain and kept in a
...

...

...

...
servers:
...
security: []
paths:
...
  /market/config:
    get:
      tags: [Market]
      summary: Fee schedule and limits
      operationId: getMarketConfig
      responses:
        '200':
          description: The `marketplace` block of `config/tokenomics.json`. Settlement is `prepaid` in this version.
          content:
            application/json:
              example: { enabled: true, feeBps: 250, feePercent: 2.5, feeToHoldersBps: 5000, minListingUsd: 1, minFillUsd: 0.01, maxDiscountBps: 7000, listingTtlHours: 168, settlement: prepaid, starterTransferable: false, creditExpiryDays: 90 }
        '404': { description: 'The marketplace is disabled (`marketplace.enabled: false`).' }
...
  /market/listings:
    get:
      tags: [Market]
      summary: Open listings, paginated
      operationId: listMarketListings
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: discountBps, in: query, schema: { type: integer, minimum: 0, maximum: 10000 }, description: 'Only this tier.' }
      responses:
        '200':
          description: Deepest discount first, then oldest first. Seller addresses are not exposed.
          content:
            application/json:
              example: { listings: [{ id: lst_6f1c2b3a4d5e6f708192, amountUsd: 100, remainingUsd: 60, soldUsd: 40, discountBps: 3000, pricePerUsd: 0.7, status: open, created_at: 1760000000, expires_at: 1760604800, closed_at: null }], total: 1, limit: 50, offset: 0 }
    post:
      tags: [Market]
      summary: List credits for sale
      description: Escrows `amountUsd` of the caller's spendable credits (ledger kind `market_escrow`) into a new listing. Unused starter credit cannot be listed (`GET /me/market → listableUsd`). Listing does not stop the 90-day clock; a buyer's credits start a fresh one. 30 writes a minute per IP.
      operationId: createMarketListing
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountUsd, discountBps]
              properties:
                amountUsd: { type: number, minimum: 1, description: 'Face value of credit to sell.' }
                discountBps: { type: integer, minimum: 0, maximum: 7000 }
      responses:
        '201':
          description: The listing plus what the seller receives if it all sells.
          content:
            application/json:
              example: { id: lst_6f1c2b3a4d5e6f708192, seller: '9xQe…', amountUsd: 100, remainingUsd: 100, soldUsd: 0, discountBps: 3000, pricePerUsd: 0.7, status: open, created_at: 1760000000, expires_at: 1760604800, closed_at: null, ifFullySold: { buyerPaysUsd: 70, feeUsd: 1.75, youReceiveUsd: 68.25 } }
        '400': { description: '`below_minimum`, `discount_too_deep`, or a malformed body.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { description: '`insufficient_credits`: spendable balance is smaller than the amount. `non_transferable`: the balance covers it only by counting starter credit, which cannot be sold.' }
  /market/listings/{id}:
    delete:
      tags: [Market]
      summary: Cancel a listing
      description: Owner only. What is left returns to the spendable balance (`market_refund`). Idempotent for a closed listing.
      operationId: cancelMarketListing
      security: [{ session: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 'The listing, now `cancelled`.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: 'No such listing for this wallet.' }
  /market/fills:
    post:
      tags: [Market]
      summary: Buy (part of) a listing
      description: Pays the discounted price from the caller's prepaid balance and credits the face value at once. Partial fills down to $0.01 (or the whole remainder). One transaction; the fee is split between the next holder pool and the treasury.
      operationId: fillMarketListing
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listingId, amountUsd]
              properties:
                listingId: { type: string }
                amountUsd: { type: number, description: 'Face value of credit to buy.' }
      responses:
        '201':
          content:
            application/json:
              example: { id: fill_2b3a4d5e6f708192a3b4, listingId: lst_6f1c2b3a4d5e6f708192, buyer: '3pQ9…', seller: '9xQe…', creditsUsd: 40, paidUsd: 28, feeUsd: 0.7, feeToHoldersUsd: 0.35, feeToTreasuryUsd: 0.35, sellerReceivedUsd: 27.3, discountBps: 3000, settlement: prepaid, settlementRef: null, created_at: 1760000100, creditBalanceUsd: 40, prepaidBalanceUsd: 22 }
        '400': { description: '`below_minimum` or a malformed body.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { description: '`insufficient_prepaid`.' }
        '404': { description: 'No such listing.' }
        '409': { description: '`own_listing`, `listing_closed` or `insufficient_depth`.' }
...
  /market/quote:
    get:
      tags: [Market]
      summary: Exact fee math for a hypothetical trade
      operationId: getMarketQuote
      parameters:
        - { name: amountUsd, in: query, required: true, schema: { type: number } }
        - { name: discountBps, in: query, required: true, schema: { type: integer } }
      responses:
        '200':
          content:
            application/json:
              example: { creditsUsd: 100, discountBps: 3000, pricePerUsd: 0.7, buyerPaysUsd: 70, feeUsd: 1.75, feeToHoldersUsd: 0.875, feeToTreasuryUsd: 0.875, sellerReceivesUsd: 68.25 }
  /me/market:
    get:
      tags: [Market]
      summary: My market activity
      description: Listings, fills as buyer and as seller, prepaid balance with its recent ledger, and withdrawal requests.
      operationId: getMyMarket
      security: [{ session: [] }]
      responses:
        '200':
          content:
            application/json:
              example: { wallet: '9xQe…', creditBalanceUsd: 10, nonTransferableUsd: 2, listableUsd: 8, prepaid: { usd: 68.25, usdMicros: 68250000, ledger: [{ id: 3, kind: market_sale, deltaUsd: 40.95, ref: 'fill:fill_…', created_at: 1760000200 }] }, listings: [], fills: { asBuyer: [], asSeller: [] }, withdrawals: [] }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/market/withdraw:
    post:
      tags: [Market]
      summary: Request a payout of prepaid USD
      description: The amount leaves the prepaid balance at once (`withdrawal` row); an operator pays it and marks the request paid.
      operationId: requestMarketWithdrawal
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [amountUsd], properties: { amountUsd: { type: number } } }
      responses:
        '201': { description: '`{ id, wallet, amountUsd, status: pending, created_at, prepaidBalanceUsd }`.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { description: '`insufficient_prepaid`.' }
...
  /admin/prepaid:
    post:
      tags: [Market]
      summary: Credit a wallet's prepaid balance
      description: 'Admin (`x-admin-token`). Records an off-chain (or hand-sent USDC) payment as prepaid balance. `ref` makes a re-post of the same payment a no-op (200, `duplicate: true`). Audited with the note.'
      operationId: adminPrepaidTopUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet, amountUsd, note]
              properties:
                wallet: { type: string }
                amountUsd: { type: number }
                note: { type: string, description: 'Where the money came from (bank ref, tx).' }
                ref: { type: string, description: "Idempotency key, e.g. the payment's own id." }
      responses:
        '201': { description: '`{ wallet, amountUsd, ref, ledgerId, duplicate: false, prepaidBalanceUsd }`.' }
        '200': { description: 'Same shape with `duplicate: true`; nothing was written.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/market/withdrawals/{id}/paid:
    post:
      tags: [Market]
      summary: Mark a withdrawal paid
      description: Admin (`x-admin-token`). Idempotent. Audited with the payout reference.
      operationId: adminMarketWithdrawalPaid
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { txRef: { type: string }, note: { type: string } } }
      responses:
        '200': { description: 'The request, now `paid`.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: 'No such request.' }
  /admin/market:
    get:
      tags: [Market]
      summary: Operator overview
      description: 'Admin (`x-admin-token`). Totals, last 24h, the open book, prepaid outstanding, the pending holder pool and every pending withdrawal.'
      operationId: adminMarketOverview
      responses:
        '200': { description: '`{ totals, last24h, book, prepaid, poolExtra, withdrawals: { pendingUsd, paidUsd, pending: [...] }, config }`.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/market/reap:
    post:
      tags: [Market]
      summary: Expire old listings now
      description: Admin (`x-admin-token`). Runs anyway once a minute and on every read of the book.
      operationId: adminMarketReap
      responses:
        '200': { description: '`{ listingsExpired }`.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 7-day session from `POST /auth/verify`.
    apiKey:
      type: http
      scheme: bearer
      description: '`mesh_sk_…` from `POST /keys`. Works wherever an OpenAI key works.'
    nodeToken:
      type: http
      scheme: bearer
      description: '`mesh_node_…` from `POST /nodes/register`.'
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: unauthorized, message: valid session token required }
...
  schemas:
...
    Error:
      type: object
      description: Error shape outside `/v1`.
      required: [error]
      properties:
        error: { type: string, description: Machine-readable code. }
        message: { type: string }
        statusCode: { type: integer }
        requestId: { type: string }
        issues: { type: array, items: { type: object } }
...